fix: rename DatabaseConnection file to DatabaseConnectionManager | [#210]
